fix: Article 로직 수정

+++ src/main/java/api/article/ArticleService.java
@@ -1,6 +1,11 @@
 package api.article;
 
 import api.article.dto.ArticleDto;
+import api.common.util.auth.loggedInUser.LoggedInUser;
+import api.user.dto.UserAccountDto;
+import api.user.enums.Role;
+import api.user.owner.Owner;
+import api.user.owner.OwnerRepository;
 import api.user.userAccount.*;
 
 
@@ -9,16 +14,17 @@ import org.springframework.stereotype.Service;
 
 import java.nio.file.AccessDeniedException;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
-    private final UserAccountRepository userAccountRepository;
+    private final OwnerRepository ownerRepository;
 
-    public ArticleService(ArticleRepository articleRepository, UserAccountRepository userAccountRepository) {
+    public ArticleService(ArticleRepository articleRepository, OwnerRepository ownerRepository) {
         this.articleRepository = articleRepository;
-        this.userAccountRepository = userAccountRepository;
+        this.ownerRepository = ownerRepository;
     }
 
     @Transactional
@@ -33,13 +39,19 @@ public class ArticleService {
     }
 
     @Transactional
-    public ArticleDto createArticle(ArticleDto articleDto) {
+    public ArticleDto createArticle(ArticleDto articleDto) throws AccessDeniedException {
+        UserAccountDto loggedInUserAccountDto = LoggedInUser.getLoggedInUserAccountDto();
+        if (loggedInUserAccountDto == null || !Role.OWNER.equals(loggedInUserAccountDto.getRole())) {
+            throw new AccessDeniedException("Only owners can create article.");
+        }
+        Optional<Owner> optionalOwner = ownerRepository.findByEmail(loggedInUserAccountDto.getEmail());
+        if (optionalOwner.isEmpty()) {
+            throw new AccessDeniedException("Owner not found for the logged in user.");
+        }
         if (articleDto.getOwnerId() == null) {
             throw new IllegalArgumentException("Owner ID must not be null");
         }
-
-        UserAccount owner = userAccountRepository.findById(articleDto.getOwnerId())
-                .orElseThrow(() -> new IllegalArgumentException("user is not found"));
+        Owner owner = optionalOwner.get();
         Article newArticle = new Article(articleDto.getTitle(), owner, articleDto.getArticleContractDetail());
         Article savedArticle = articleRepository.save(newArticle);
 
@@ -48,15 +60,51 @@ public class ArticleService {
 
 
     @Transactional
-    public ArticleDto updateArticle(Integer id, ArticleDto articleDto) {
+    public ArticleDto updateArticle(Integer id, ArticleDto articleDto) throws AccessDeniedException {
+        UserAccountDto loggedInUserAccountDto = LoggedInUser.getLoggedInUserAccountDto();
+        if (loggedInUserAccountDto == null || !Role.OWNER.equals(loggedInUserAccountDto.getRole())) {
+            throw new AccessDeniedException("Only owners can update article.");
+        }
+
+        Optional<Owner> optionalOwner = ownerRepository.findByEmail(loggedInUserAccountDto.getEmail());
+        if (optionalOwner.isEmpty()) {
+            throw new AccessDeniedException("Owner not found for the logged in user.");
+        }
+
+        if (articleDto.getOwnerId() == null) {
+            throw new IllegalArgumentException("Owner ID must not be null");
+        }
+
+        Owner owner = optionalOwner.get();
         Article article = articleRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid article ID"));
+
+        if (!article.getOwner().equals(owner)) {
+            throw new AccessDeniedException("Only the author can update their own post.");
+        }
+
         Article updatedArticle = article.update(articleDto.getTitle(), articleDto.getArticleContractDetail());
         Article savedArticle = articleRepository.save(updatedArticle);
         return new ArticleDto(savedArticle);
     }
 
     @Transactional
-    public void deleteArticle(Integer id) {
+    public void deleteArticle(Integer id) throws AccessDeniedException {
+        UserAccountDto loggedInUserAccountDto = LoggedInUser.getLoggedInUserAccountDto();
+        if (loggedInUserAccountDto == null || !Role.OWNER.equals(loggedInUserAccountDto.getRole())) {
+            throw new AccessDeniedException("Only owners can update article.");
+        }
+
+        Optional<Owner> optionalOwner = ownerRepository.findByEmail(loggedInUserAccountDto.getEmail());
+        if (optionalOwner.isEmpty()) {
+            throw new AccessDeniedException("Owner not found for the logged in user.");
+        }
+
+        Owner owner = optionalOwner.get();
+        Article article = articleRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("Invalid article ID"));
+
+        if (!article.getOwner().equals(owner)) {
+            throw new AccessDeniedException("Only the author can delete their own post.");
+        }
         articleRepository.deleteById(id);
     }
